Refactor Impression commands up to EditImpressionCommand
