+++ src/main/java/duke/command/impression/ImpressionDeleteCommand.java
@@ -16,18 +16,5 @@ public class ImpressionDeleteCommand extends ArgCommand {
         return ImpressionDeleteSpec.getSpec();
     }
 
-    @Override
-    public void execute(DukeCore core) throws DukeException {
-        Impression impression = ImpressionUtils.getImpression(core);
-        DukeData delData = ImpressionUtils.getData(getArg(), getSwitchVal("evidence"),
-                getSwitchVal("treatment"), impression);
-        String delMsg = "'" + delData.getName() + "' deleted!";
-        if (delData instanceof Evidence) {
-            impression.deleteEvidence(delData.getName());
-        } else if (delData instanceof Treatment) {
-            impression.deleteTreatment(delData.getName());
-        }
-        core.writeJsonFile();
-        core.updateUi(delMsg);
-    }
+
 }
