+++ src/main/java/duke/command/impression/ImpressionDeleteSpec.java
@@ -1,8 +1,15 @@
 package duke.command.impression;
 
+import duke.DukeCore;
+import duke.command.ArgCommand;
 import duke.command.ArgLevel;
 import duke.command.ArgSpec;
 import duke.command.Switch;
+import duke.data.DukeData;
+import duke.data.Evidence;
+import duke.data.Impression;
+import duke.data.Treatment;
+import duke.exception.DukeException;
 
 public class ImpressionDeleteSpec extends ArgSpec {
     private static final ImpressionDeleteSpec spec = new ImpressionDeleteSpec();
@@ -17,5 +24,20 @@ public class ImpressionDeleteSpec extends ArgSpec {
                 new Switch("evidence", String.class, true, ArgLevel.REQUIRED, "e"),
                 new Switch("treatment", String.class, true, ArgLevel.REQUIRED, "t")
         );
+    }
+
+    @Override
+    public void execute(DukeCore core, ArgCommand cmd) throws DukeException {
+        Impression impression = ImpressionUtils.getImpression(core);
+        DukeData delData = ImpressionUtils.getData(cmd.getArg(), cmd.getSwitchVal("evidence"),
+                cmd.getSwitchVal("treatment"), impression);
+        String delMsg = "'" + delData.getName() + "' deleted!";
+        if (delData instanceof Evidence) {
+            impression.deleteEvidence(delData.getName());
+        } else if (delData instanceof Treatment) {
+            impression.deleteTreatment(delData.getName());
+        }
+        core.writeJsonFile();
+        core.updateUi(delMsg);
     }
 }
