+++ src/main/java/duke/command/impression/ImpressionEditSpec.java
@@ -1,16 +1,32 @@
 package duke.command.impression;
 
+import duke.DukeCore;
+import duke.command.ArgCommand;
 import duke.command.ArgLevel;
-import duke.command.ArgSpec;
 import duke.command.Switch;
+import duke.data.DukeData;
+import duke.data.Impression;
+import duke.data.Investigation;
+import duke.data.Medicine;
+import duke.data.Observation;
+import duke.data.Patient;
+import duke.data.Plan;
+import duke.data.Result;
+import duke.data.Treatment;
+import duke.exception.DukeException;
+import duke.exception.DukeHelpException;
+import duke.exception.DukeUtilException;
 
-public class ImpressionEditSpec extends ArgSpec {
+import java.util.List;
+import java.util.Map;
+
+public class ImpressionEditSpec extends DukeDataSpec {
 
     private static final ImpressionEditSpec spec = new ImpressionEditSpec();
-
     public static ImpressionEditSpec getSpec() {
         return spec;
     }
+    private ArgCommand cmd;
 
     private ImpressionEditSpec() {
         cmdArgLevel = ArgLevel.NONE;
@@ -32,5 +48,161 @@ public class ImpressionEditSpec extends ArgSpec {
                 new Switch("objective", null, true, ArgLevel.NONE, "obj"),
                 new Switch("description", String.class, true, ArgLevel.REQUIRED, "desc")
         );
+    }
+
+    @Override
+    public void execute(DukeCore core, ArgCommand cmd) throws DukeException {
+        String editType = uniqueDataType(cmd);
+        DukeData editData;
+        boolean isAppending = false;
+        if (cmd.isSwitchSet("append")) {
+            isAppending = true;
+        }
+
+        if (editType == null) { // edit impression
+            editImpression(ImpressionUtils.getImpression(core), isAppending);
+        } else {
+            if (cmd.isSwitchSet("description")) {
+                throw new DukeHelpException("Descriptions are only for impressions!", cmd);
+            }
+
+            // find DukeData of editType
+            // TODO: proper search with disambiguation
+            // TODO: select by index
+
+            String editDataName = cmd.getSwitchVal(editType);
+            List<DukeData> resultList = ImpressionUtils.getImpression(core).find(editDataName);
+            editData = findDataOfClass(resultList, getDataClass(editType));
+
+            if (editData == null) {
+                throw new DukeHelpException("I can't find a " + editType + " item named '" + editDataName + "'!", cmd);
+            }
+
+            // process status
+            switch (editType) {
+            case "plan":
+                editStatus(editData, Plan.getStatusArr());
+                break;
+            case "medicine":
+                editStatus(editData, Medicine.getStatusArr());
+                break;
+            case "investigation":
+                editStatus(editData, Investigation.getStatusArr());
+                break;
+            default:
+                break;
+            }
+
+            // process remaining switches entered
+            for (Map.Entry<String, String> entry : cmd.getSwitchVals().entrySet()) {
+                String switchName = entry.getKey();
+                String entryStr = entry.getValue();
+                int entryInt = 0;
+                if (Integer.class.equals(getSwitchMap().get(switchName).type)) {
+                    entryInt = cmd.switchToInt(switchName);
+                }
+
+                // ignore switches that don't need processing
+                if (switchName.equals(editType)) {
+                    continue;
+                }
+
+                // TODO: deal with null
+                switch (switchName) {
+                case "name":
+                    //TODO check for allergies for medicine
+                    editData.setName((isAppending) ? editData.getName() + entryStr : entryStr);
+                    break;
+                case "priority":
+                    if (entryInt == -1) {
+                        break;
+                    }
+                    ImpressionUtils.checkPriority(entryInt);
+                    editData.setPriority(entryInt);
+                    break;
+                case "summary":
+                    editData.setSummary((isAppending) ? editData.getSummary() + entryStr : entryStr);
+                    break;
+                default:
+                    switch (editType) {
+                    case "medicine":
+                        Medicine med = (Medicine) editData;
+                        switch (switchName) {
+                        case "dose":
+                            med.setDose((isAppending) ? med.getDose() + entryStr : entryStr);
+                            break;
+                        case "date":
+                            med.setStartDate((isAppending) ? med.getStartDate() + entryStr : entryStr);
+                            break;
+                        case "duration":
+                            med.setDuration((isAppending) ? med.getDuration() + entryStr : entryStr);
+                            break;
+                        default:
+                            throw new DukeHelpException("Medicine plans do not have this property: '"
+                                    + entryStr + "'", this);
+                        }
+                        break;
+                    case "observation":
+                        Observation obsv = (Observation) editData;
+                        if (isSwitchSet("objective") && isSwitchSet("subjective")) {
+                            throw new DukeHelpException("I don't know if you want this observation to be objective"
+                                    + "or subjective!", this);
+                        } else if (isSwitchSet("objective")) {
+                            obsv.setObjective(true);
+                        } else if (isSwitchSet("subjective")) {
+                            obsv.setObjective(false);
+                        }
+                        break;
+                    case "plan": //fallthrough
+                    case "result": //fallthrough
+                    case "investigation": //all switches should already be handled
+                        break;
+                    default:
+                        throw new DukeException("Invalid data type found when making edits!");
+                    }
+                    break;
+                }
+
+            }
+        }
+
+        core.writeJsonFile();
+        core.updateUi("Details updated");
+    }
+
+    private void editStatus(DukeData editData, List<String> statusList) throws DukeUtilException {
+        String statusStr = cmd.getSwitchVal("status");
+        if (statusStr != null) {
+            Treatment treatment = (Treatment) editData;
+            treatment.setStatusIdx(ImpressionUtils.processStatus(statusStr, statusList));
+        }
+    }
+
+    private void editImpression(Impression impression, boolean isAppending) {
+        Patient patient = (Patient) impression.getParent();
+
+        String newName = getSwitchVal("name");
+        if (newName != null) {
+            patient.getImpressionsObservableMap().remove(impression.getName());
+            impression.setName((isAppending) ? impression.getName() + newName : newName);
+            patient.getImpressionsObservableMap().put(newName, impression);
+        }
+
+        String newDesc = getSwitchVal("description");
+        if (newDesc != null) {
+            impression.setDescription((isAppending) ? impression.getDescription() + newDesc : newDesc);
+        }
+    }
+
+    // TODO proper searching
+    // simply gets the first search result with a matching class
+    // this is a terrible method, kill it and replace it with proper disambiguation as soon as possible
+    private DukeData findDataOfClass(List<DukeData> resultList, Class dataClass) {
+        for (DukeData result : resultList) {
+            if (result.getClass() == dataClass) {
+                return result;
+            }
+        }
+        return null;
     }
 }
